+++ src/cloudant/database.py
@@ -17,7 +17,6 @@ API module that maps to a Cloudant or CouchDB database instance.
 """
 import json
 import contextlib
-import posixpath
 
 from requests.exceptions import HTTPError
 
@@ -84,10 +83,8 @@ class CouchDatabase(dict):
 
         :returns: Database URL
         """
-        return posixpath.join(
-            self._database_host,
-            url_quote_plus(self.database_name)
-        )
+        return '/'.join((
+            self._database_host, url_quote_plus(self.database_name)))
 
     @property
     def creds(self):
@@ -189,7 +186,7 @@ class CouchDatabase(dict):
 
         :returns: All design documents found in this database in JSON format
         """
-        url = posixpath.join(self.database_url, '_all_docs')
+        url = '/'.join((self.database_url, '_all_docs'))
         query = "startkey=\"_design\"&endkey=\"_design0\"&include_docs=true"
         resp = self.r_session.get(url, params=query)
         resp.raise_for_status()
@@ -203,7 +200,7 @@ class CouchDatabase(dict):
 
         :returns: List of names for all design documents in this database
         """
-        url = posixpath.join(self.database_url, '_all_docs')
+        url = '/'.join((self.database_url, '_all_docs'))
         query = "startkey=\"_design\"&endkey=\"_design0\""
         resp = self.r_session.get(url, params=query)
         resp.raise_for_status()
@@ -631,18 +628,20 @@ class CouchDatabase(dict):
         if not remote:
             super(CouchDatabase, self).__iter__()
         else:
-            next_startkey = '0'
+            # Use unicode Null U+0000 as the initial lower bound to ensure any
+            # document id could exist in the results set.
+            next_startkey = u'\u0000'
             while next_startkey is not None:
                 docs = self.all_docs(
-                    limit=self._fetch_limit + 1,  # Get one extra doc
-                                                  # to use as
-                                                  # next_startkey
+                    limit=self._fetch_limit,
                     include_docs=True,
                     startkey=next_startkey
                 ).get('rows', [])
 
-                if len(docs) > self._fetch_limit:
-                    next_startkey = docs.pop()['id']
+                if len(docs) >= self._fetch_limit:
+                    # Ensure the next document batch contains ids that sort
+                    # strictly higher than the previous document id fetched.
+                    next_startkey = docs[-1]['id'] + u'\u0000'
                 else:
                     # This is the last batch of docs, so we set
                     # ourselves up to break out of the while loop
@@ -674,7 +673,7 @@ class CouchDatabase(dict):
 
         :returns: Bulk document creation/update status in JSON format
         """
-        url = posixpath.join(self.database_url, '_bulk_docs')
+        url = '/'.join((self.database_url, '_bulk_docs'))
         data = {'docs': docs}
         headers = {'Content-Type': 'application/json'}
         resp = self.r_session.post(
@@ -697,7 +696,7 @@ class CouchDatabase(dict):
 
         :returns: List of missing document revision values
         """
-        url = posixpath.join(self.database_url, '_missing_revs')
+        url = '/'.join((self.database_url, '_missing_revs'))
         data = {doc_id: list(revisions)}
 
         resp = self.r_session.post(
@@ -726,7 +725,7 @@ class CouchDatabase(dict):
 
         :returns: The revision differences in JSON format
         """
-        url = posixpath.join(self.database_url, '_revs_diff')
+        url = '/'.join((self.database_url, '_revs_diff'))
         data = {doc_id: list(revisions)}
 
         resp = self.r_session.post(
@@ -745,7 +744,7 @@ class CouchDatabase(dict):
 
         :returns: Revision limit value for the current remote database
         """
-        url = posixpath.join(self.database_url, '_revs_limit')
+        url = '/'.join((self.database_url, '_revs_limit'))
         resp = self.r_session.get(url)
         resp.raise_for_status()
 
@@ -766,7 +765,7 @@ class CouchDatabase(dict):
 
         :returns: Revision limit set operation status in JSON format
         """
-        url = posixpath.join(self.database_url, '_revs_limit')
+        url = '/'.join((self.database_url, '_revs_limit'))
 
         resp = self.r_session.put(url, data=json.dumps(limit, cls=self.client.encoder))
         resp.raise_for_status()
@@ -780,7 +779,7 @@ class CouchDatabase(dict):
 
         :returns: View cleanup status in JSON format
         """
-        url = posixpath.join(self.database_url, '_view_cleanup')
+        url = '/'.join((self.database_url, '_view_cleanup'))
         resp = self.r_session.post(
             url,
             headers={'Content-Type': 'application/json'}
@@ -957,8 +956,8 @@ class CloudantDatabase(CouchDatabase):
 
         :returns: Security document URL
         """
-        parts = ['_api', 'v2', 'db', self.database_name, '_security']
-        url = posixpath.join(self._database_host, *parts)
+        url = '/'.join((self._database_host, '_api', 'v2', 'db',
+                        self.database_name, '_security'))
         return url
 
     def share_database(self, username, roles=None):
@@ -1039,7 +1038,7 @@ class CloudantDatabase(CouchDatabase):
 
         :returns: Shard information retrieval status in JSON format
         """
-        url = posixpath.join(self.database_url, '_shards')
+        url = '/'.join((self.database_url, '_shards'))
         resp = self.r_session.get(url)
         resp.raise_for_status()
 
@@ -1058,7 +1057,7 @@ class CloudantDatabase(CouchDatabase):
         :returns: The query indexes in the database
         """
 
-        url = posixpath.join(self.database_url, '_index')
+        url = '/'.join((self.database_url, '_index'))
         resp = self.r_session.get(url)
         resp.raise_for_status()