+++ tests/unit/database_tests.py
@@ -25,10 +25,10 @@ module docstring.
 import unittest
 import mock
 import requests
-import posixpath
 import os
 import uuid
 
+from cloudant._2to3 import UNICHR
 from cloudant.result import Result, QueryResult
 from cloudant.error import CloudantArgumentError, CloudantDatabaseException
 from cloudant.document import Document
@@ -148,8 +148,8 @@ class DatabaseTests(UnitTestDbBase):
         """
         self.assertEqual(
             self.db.database_url,
-            posixpath.join(self.client.server_url, self.test_dbname)
-            )
+            '/'.join((self.client.server_url, self.test_dbname))
+        )
 
     def test_retrieve_creds(self):
         """
@@ -232,8 +232,7 @@ class DatabaseTests(UnitTestDbBase):
         same.  Therefore comparing keys is a valid test of this functionality.
         """
         resp = self.db.r_session.get(
-            posixpath.join(self.client.server_url, self.test_dbname)
-            )
+            '/'.join((self.client.server_url, self.test_dbname)))
         expected = resp.json()
         actual = self.db.metadata()
         self.assertListEqual(list(actual.keys()), list(expected.keys()))
@@ -605,6 +604,39 @@ class DatabaseTests(UnitTestDbBase):
             self.assertEqual(doc['name'], 'julia')
             self.assertEqual(doc['age'], int(id[len(id) - 3: len(id)]))
 
+    def test_document_iteration_completeness(self):
+        """
+        Test __iter__ works as expected, fetching all documents from the
+        database.
+        """
+        for _ in self.db:
+            self.fail('There should be no documents in the database yet!!')
+
+        # sample code point ranges
+        include_ranges = [
+            (0x0023, 0x0026),
+            (0x00A1, 0x00AC),
+            (0x0370, 0x0377),
+            (0x037A, 0x037E),
+            (0x0384, 0x038A),
+            (0x16A0, 0x16F0),
+            (0x2C60, 0x2C7F)
+        ]
+
+        all_docs = [{'_id': UNICHR(i) + UNICHR(j)} for a, b in include_ranges
+                                                   for i in range(a, b)
+                                                   for j in range(a, b)]
+        batch_size = 500
+        for i in range(0, len(all_docs), batch_size):
+            self.db.bulk_docs(all_docs[i:i+batch_size])
+
+        doc_count = 0
+        for i, doc in enumerate(self.db):
+            doc_count += 1
+            self.assertEqual(doc['_id'], all_docs[i]['_id'])
+
+        self.assertEqual(doc_count, len(all_docs))
+
     def test_document_iteration_returns_valid_documents(self):
         """
         This test will check that the __iter__ method returns documents that are
@@ -622,7 +654,7 @@ class DatabaseTests(UnitTestDbBase):
             # A valid document must have a document_url
             self.assertEqual(
                 doc.document_url,
-                posixpath.join(self.db.database_url, doc['_id'])
+                '/'.join((self.db.database_url, doc['_id']))
             )
             if isinstance(doc, DesignDocument):
                 self.assertEqual(doc['_id'], '_design/ddoc001')